refactor: extract shared S3 fetch into pkg/handler
The metricsconfigurator and pollerconfigurator subpackages each carried
their own copy of the same plumbing: a parseS3URI helper that splits an
s3:// URI into bucket+key, and a downloadX function that called
s3.GetObject and read the body. The two diverged only in how they
interpreted the bytes (YAML for metric stream filters, JSON for poller
config).

Hoist the S3 plumbing into pkg/handler:

- pkg/handler/s3.go exposes ParseS3URI(uri) (bucket, key, err) and
  GetS3Object(ctx, awsCfg, uri) ([]byte, error). GetS3Object returns
  raw bytes; the caller decides how to interpret them.
- pkg/handler/s3_test.go covers ParseS3URI (the duplicated tests merged
  into one).
- metricsconfigurator/filteruri.go drops parseS3URI and
  downloadFilterYAML; handler.go calls handler.GetS3Object directly and
  hands the bytes to parseFilterYAML, which is now the only thing in
  filteruri.go besides the YAML schema types.
- pollerconfigurator/s3config.go drops parseS3URI and the
  downloadPollerConfigFromS3 wrapper; downloadPollerConfig now uses
  handler.GetS3Object on s3:// URIs and falls through to
  downloadPollerConfigFromURL for http(s):// URIs (which the tests use
  with httptest servers). parsePollerConfig still takes io.Reader for
  the URL path; the S3 path wraps the bytes in bytes.NewReader.
- The duplicate parseS3URI tests in both subpackages are removed; the
  canonical test now lives next to the function.

Net effect: -184 / +123 lines.
